move tray initialization to the base class close_loop_env. add close_loop_tray parameter to include the tray in close loop envs. change default view_scale to 1.5
